redistributeLengths coded, tested

+++ src/composer/Chord.java
@@ -2,13 +2,13 @@ package composer;
 
 public class Chord {
 	public String name; //name of chord. note: "N" indicates no change of chord in unimportant locations within the music.
-	public int length; //number of beats
-	public Chord(String name, int length) {
+	public double length; //number of beats
+	public Chord(String name, double length) {
 		this.name = name;
 		this.length = length;
 	}
 	public String getName() { return name; }
-	public int getLength() { return length; }
+	public double getLength() { return length; }
 	public void setName(String name) { this.name = name; }
-	public void setLength(int length) { this.length = length; }
+	public void setLength(double length) { this.length = length; }
 }
